web: use relative imports
