+++ web/src/App.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import {Fabric} from 'office-ui-fabric-react/lib/Fabric';
+import { Fabric } from 'office-ui-fabric-react/lib/Fabric';
 import { ConnectedRouter } from 'connected-react-router';
-import {Switch, Route} from "react-router-dom";
+import { Switch, Route } from "react-router-dom";
 
 import { configureStore, createGoConsoleAdapter } from './store';
 import { history } from './store/configure';
 import { bootstrapGo } from './services/go';
-import Playground from './Playground';
+import Playground from '~/components/pages/Playground';
 import './App.css';
 import config from './services/config'
 
@@ -19,20 +19,20 @@ config.sync();
 bootstrapGo(createGoConsoleAdapter(a => store.dispatch(a)));
 
 function App() {
-  return (
-    <Provider store={store}>
-        <ConnectedRouter history={history}>
-            <Fabric className="App">
-                <Switch>
-                    <Route
-                        path="/(snippet)?/:snippetID?"
-                        component={Playground}
-                    />
-                </Switch>
-            </Fabric>
-        </ConnectedRouter>
-    </Provider>
-  );
+    return (
+        <Provider store={store}>
+            <ConnectedRouter history={history}>
+                <Fabric className="App">
+                    <Switch>
+                        <Route
+                            path="/(snippet)?/:snippetID?"
+                            component={Playground}
+                        />
+                    </Switch>
+                </Fabric>
+            </ConnectedRouter>
+        </Provider>
+    );
 }
 
 export default App;
