+++ web/src/components/core/Header.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
-import './Header.css'
 import { MessageBarButton } from 'office-ui-fabric-react';
-import {CommandBar, ICommandBarItemProps} from 'office-ui-fabric-react/lib/CommandBar';
+import { CommandBar, ICommandBarItemProps } from 'office-ui-fabric-react/lib/CommandBar';
 import { MessageBar, MessageBarType } from 'office-ui-fabric-react/lib/MessageBar';
-import {getTheme} from '@uifabric/styling';
-import SettingsModal, {SettingsChanges} from './settings/SettingsModal';
-import AboutModal from './AboutModal';
-import config from './services/config';
-import api from './services/api';
-import { getSnippetsMenuItems, SnippetMenuItem } from './utils/headerutils';
+import { getTheme } from '@uifabric/styling';
+import SettingsModal, { SettingsChanges } from '~/settings/SettingsModal';
+import AboutModal from '~/components/modals/AboutModal';
+import config from '~/services/config';
+import api from '~/services/api';
+import { getSnippetsMenuItems, SnippetMenuItem } from '~/utils/headerutils';
 import {
     Connect,
     dispatchToggleTheme,
@@ -19,8 +18,9 @@ import {
     runFileDispatcher,
     saveFileDispatcher,
     shareSnippetDispatcher
-} from './store';
-import ChangeLogModal from "./ChangeLogModal";
+} from '~/store';
+import ChangeLogModal from '@components/modals/ChangeLogModal';
+import './Header.css';
 
 interface HeaderState {
     showSettings: boolean
@@ -30,7 +30,7 @@ interface HeaderState {
     showUpdateBanner: boolean
 }
 
-@Connect(s => ({darkMode: s.settings.darkMode, loading: s.status?.loading}))
+@Connect(s => ({ darkMode: s.settings.darkMode, loading: s.status?.loading }))
 export class Header extends React.Component<any, HeaderState> {
     private fileInput?: HTMLInputElement;
     private snippetMenuItems = getSnippetsMenuItems(i => this.onSnippetMenuItemClick(i));
@@ -55,9 +55,9 @@ export class Header extends React.Component<any, HeaderState> {
 
         // show update popover
         api.getVersion().then(r => {
-            const {version} = r;
+            const { version } = r;
             if (!version) return;
-            this.setState({showUpdateBanner: version !== config.appVersion});
+            this.setState({ showUpdateBanner: version !== config.appVersion });
         }).catch(err => console.warn('failed to check server API version: ', err));
     }
 
@@ -81,7 +81,7 @@ export class Header extends React.Component<any, HeaderState> {
                 key: 'openFile',
                 text: 'Open',
                 split: true,
-                iconProps: {iconName: 'OpenFile'},
+                iconProps: { iconName: 'OpenFile' },
                 disabled: this.props.loading,
                 onClick: () => this.fileInput?.click(),
                 subMenuProps: {
@@ -93,7 +93,7 @@ export class Header extends React.Component<any, HeaderState> {
                 text: 'Run',
                 ariaLabel: 'Run program (Ctrl+Enter)',
                 title: 'Run program (Ctrl+Enter)',
-                iconProps: {iconName: 'Play'},
+                iconProps: { iconName: 'Play' },
                 disabled: this.props.loading,
                 onClick: () => {
                     this.props.dispatch(runFileDispatcher);
@@ -102,7 +102,7 @@ export class Header extends React.Component<any, HeaderState> {
             {
                 key: 'share',
                 text: 'Share',
-                iconProps: {iconName: 'Share'},
+                iconProps: { iconName: 'Share' },
                 disabled: this.props.loading,
                 onClick: () => {
                     this.props.dispatch(shareSnippetDispatcher);
@@ -111,7 +111,7 @@ export class Header extends React.Component<any, HeaderState> {
             {
                 key: 'download',
                 text: 'Download',
-                iconProps: {iconName: 'Download'},
+                iconProps: { iconName: 'Download' },
                 disabled: this.props.loading,
                 onClick: () => {
                     this.props.dispatch(saveFileDispatcher);
@@ -121,10 +121,10 @@ export class Header extends React.Component<any, HeaderState> {
                 key: 'settings',
                 text: 'Settings',
                 ariaLabel: 'Settings',
-                iconProps: {iconName: 'Settings'},
+                iconProps: { iconName: 'Settings' },
                 disabled: this.props.loading,
                 onClick: () => {
-                    this.setState({showSettings: true});
+                    this.setState({ showSettings: true });
                 }
             }
         ];
@@ -137,9 +137,9 @@ export class Header extends React.Component<any, HeaderState> {
                 text: 'What\'s new',
                 ariaLabel: 'Changelog',
                 disabled: this.props.loading,
-                iconProps: {iconName: 'Giftbox'},
+                iconProps: { iconName: 'Giftbox' },
                 onClick: () => {
-                    this.setState({showChangelog: true});
+                    this.setState({ showChangelog: true });
                 }
             },
             {
@@ -148,7 +148,7 @@ export class Header extends React.Component<any, HeaderState> {
                 ariaLabel: 'Format Code (Ctrl+Shift+F)',
                 iconOnly: true,
                 disabled: this.props.loading,
-                iconProps: {iconName: 'Code'},
+                iconProps: { iconName: 'Code' },
                 onClick: () => {
                     this.props.dispatch(formatFileDispatcher);
                 }
@@ -158,7 +158,7 @@ export class Header extends React.Component<any, HeaderState> {
                 text: 'Toggle Dark Mode',
                 ariaLabel: 'Toggle Dark Mode',
                 iconOnly: true,
-                iconProps: {iconName: this.props.darkMode ? 'Brightness' : 'ClearNight'},
+                iconProps: { iconName: this.props.darkMode ? 'Brightness' : 'ClearNight' },
                 onClick: () => {
                     this.props.dispatch(dispatchToggleTheme)
                 },
@@ -172,23 +172,23 @@ export class Header extends React.Component<any, HeaderState> {
                 key: 'new-issue',
                 text: 'Submit Issue',
                 ariaLabel: 'Submit Issue',
-                iconProps: {iconName: 'Bug'},
+                iconProps: { iconName: 'Bug' },
                 onClick: () => window.open(config.issueUrl, '_blank')
             },
             {
                 key: 'donate',
                 text: 'Donate',
                 ariaLabel: 'Donate',
-                iconProps: {iconName: 'Heart'},
+                iconProps: { iconName: 'Heart' },
                 onClick: () => window.open(config.donateUrl, '_blank')
             },
             {
                 key: 'about',
                 text: 'About',
                 ariaLabel: 'About',
-                iconProps: {iconName: 'Info'},
+                iconProps: { iconName: 'Info' },
                 onClick: () => {
-                    this.setState({showAbout: true});
+                    this.setState({ showAbout: true });
                 }
             }
         ]
@@ -214,7 +214,7 @@ export class Header extends React.Component<any, HeaderState> {
             this.props.dispatch(newBuildParamsChangeDispatcher(runtime, autoFormat));
         }
 
-        this.setState({showSettings: false});
+        this.setState({ showSettings: false });
     }
 
     render() {
@@ -222,13 +222,13 @@ export class Header extends React.Component<any, HeaderState> {
             <MessageBar
                 className={this.state.showUpdateBanner ? 'app__update app__update--visible' : 'app__update'}
                 messageBarType={MessageBarType.warning}
-                onDismiss={() => this.setState({showUpdateBanner: false})}
+                onDismiss={() => this.setState({ showUpdateBanner: false })}
                 dismissButtonAriaLabel="Close"
                 isMultiline={false}
                 actions={
                     <div>
                         <MessageBarButton onClick={() => {
-                            this.setState({showUpdateBanner: false});
+                            this.setState({ showUpdateBanner: false });
                             config.forceRefreshPage();
                         }}>Action</MessageBarButton>
                     </div>
@@ -249,8 +249,8 @@ export class Header extends React.Component<any, HeaderState> {
                 ariaLabel='CodeEditor menu'
             />
             <SettingsModal onClose={(args) => this.onSettingsClose(args)} isOpen={this.state.showSettings} />
-            <AboutModal onClose={() => this.setState({showAbout: false})} isOpen={this.state.showAbout} />
-            <ChangeLogModal onClose={() => this.setState({showChangelog: false})} isOpen={this.state.showChangelog} />
+            <AboutModal onClose={() => this.setState({ showAbout: false })} isOpen={this.state.showAbout} />
+            <ChangeLogModal onClose={() => this.setState({ showChangelog: false })} isOpen={this.state.showChangelog} />
         </header>;
     }
 }
