+++ web/src/components/preview/Preview.css
@@ -1,8 +1,6 @@
 .app-preview {
-    /*background: #222;*/
-    /*color: #ccc;*/
-    max-height: 640px;
-    height: 50%;
+    /* max-height: 640px;
+    height: 50%; */
     box-sizing: border-box;
     overflow: auto;
 }
