+++ web/tsconfig.paths.json
@@ -10,6 +10,9 @@
       ],
       "@store/*": [
         "./store/*"
+      ],
+      "~/*": [
+        "./*"
       ]
     }
   }
